added isolated LocationServiceTest (with LocationDao Mock)

+++ src/main/java/sepm/dsa/service/LocationServiceImpl.java
@@ -6,6 +6,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sepm.dsa.dao.LocationDao;
+import sepm.dsa.exceptions.DSAModelNotFoundException;
 import sepm.dsa.exceptions.DSAValidationException;
 import sepm.dsa.model.Location;
 
@@ -53,10 +54,9 @@ public class LocationServiceImpl implements LocationService {
     @Transactional(readOnly = false)
     public void remove(Location location) {
         log.debug("calling remove(" + location + ")");
-//        List<Trader> tradersInTown = traderService.getAllForLocation(location);
-
-//        tradersInTown.forEach(traderService::remove);
-
+        if (location.getId() == null) {
+            throw new DSAValidationException("Ort muss eine ID haben");
+        }
         locationDao.remove(location);
     }
